Fix memory corruption, extend 'Categorizer'.
The memory corruption occurred when a 'Categorizer' object was copied or
moved, as the address of the buffer in the 'std::string' object to which
all the 'std::string_views' point can change in this case.

Note that this stands not only for copies but also for moves: moving a
string does not guarantee that its internal buffer will remain at the
same memory address.